requirements.txt
pandas==2.2.1
supabase==2.3.5
python-dotenv==1.0.1 